Move different tasks (cluster, index) in specific files.
No code change. Only move.
